Added animations to route and header

+++ src/app/user/auth.service.ts
@@ -36,26 +36,28 @@ export class AuthService {
     private actionService: ActionService,
     ) {
       const loggedInCookie = this.cookieService.get('LoggedIn');
-      this.userService.getUser(loggedInCookie).subscribe(current => {
-        if (current) {
-          this.loggedIn.next(loggedInCookie !== '');
-          this.userService.CurrentUser = current;
-          this.admin.next(this.userService.isAdmin);
-          this.pending.next(this.userService.isPending);
-          if (this.redirectUrl) {
-            this.router.navigate([this.redirectUrl]);
+      if (loggedInCookie) {
+        this.userService.getUser(loggedInCookie).subscribe(current => {
+          if (current) {
+            this.loggedIn.next(loggedInCookie !== '');
+            this.userService.CurrentUser = current;
+            this.admin.next(this.userService.isAdmin);
+            this.pending.next(this.userService.isPending);
+            if (this.redirectUrl) {
+              this.router.navigate([this.redirectUrl]);
+            }
+            this.configService.getConfig('auto-logout').subscribe(autoLogout => {
+              this.cookieService.delete('LoggedIn');
+              const expirationDate = new Date();
+              expirationDate.setMinutes(expirationDate.getMinutes() + Number(autoLogout.value));
+              this.cookieService.set('LoggedIn', loggedInCookie, expirationDate);
+              setTimeout(() => {
+                this.logout();
+              }, Number(autoLogout.value) * 60000);
+            });
           }
-          this.configService.getConfig('auto-logout').subscribe(autoLogout => {
-            this.cookieService.delete('LoggedIn');
-            const expirationDate = new Date();
-            expirationDate.setMinutes(expirationDate.getMinutes() + Number(autoLogout.value));
-            this.cookieService.set('LoggedIn', loggedInCookie, expirationDate);
-            setTimeout(() => {
-              this.logout();
-            }, Number(autoLogout.value) * 60000);
-          });
-        }
-      });
+        });
+      }
   }
 
   googleLogin() {
